+++ src/components/action-toolbar/action-toolbar-button/__definition__.js
@@ -1,0 +1,7 @@
+import ActionToolbarButton from './';
+import Definition from './../../../../demo/utils/definition';
+
+let definition = new Definition('action-toolbar-button', ActionToolbarButton, {
+});
+
+export default definition;