+++ src/components/action-toolbar/action-toolbar-multi-action-button/__definition__.js
@@ -1,0 +1,7 @@
+import ActionToolbarMultiActionButton from './';
+import Definition from './../../../../demo/utils/definition';
+
+let definition = new Definition('action-toolbar-multi-action-button', ActionToolbarMultiActionButton, {
+});
+
+export default definition;